Made BamAlignment flag queries symmetrical
 * For example: there is now a SetIsMapped() setter to match the
IsMapped() getter.  Before you had to reverse your logic on a few of the
flags (in this case, using SetIsUnmapped()).  Not impossible to use, but
not immediately obvious and intuitive, and hard to remember when to use
the opposite setter.  These older methods will remain available, but
should be considered deprecated.